Use After=network.target
Primarily because otherwise the network may go down before charon gets
the opportunity to send deletes on exit, and it matches what the old
init script did.

Note that in Debian, network.target apparently runs after ifup so the
network is in most cases really up, so we don't need
network-online.target. (And charon uses rtnetlink to get notified of
changes anyway.)