Fix handling of ipsec name under systemd (#781209)
We used to make the old ipsec name, which corresponds to the sysvinit
init script, available via an Alias directive in the systemd service
file (which follows upstream and is named strongswan.service).

Unfortunately this results in a situation where the strongswan-starter
postinst calls invoke-rc.d using the ipsec name before the service has
been installed, and thus uses the sysvinit script instead of systemd's
native unit. This results in some confusion later when the unit gets
installed.

To avoid this, ship a symlink in the package, this way the ipsec name is
available as soon as the package has been unpacked. Thanks to Michael
Biebl for the suggestion.